Update Main.java
Program is currently running but is missing a few key things for user ease. User input validation for the input integer is also causing the program to crash, despite having a try/catch put in place to allow the user to continue. Same method used for input string, which works just fine.